Deep clean of source code
