+++ jungfrau_gui/ui_components/tem_controls/gaussian_fitter.py
@@ -3,7 +3,7 @@ import numpy as np
 from PySide6.QtCore import QObject, Signal, Slot
 # from line_profiler import LineProfiler
 
-from .toolbox.fit_beam_intensity import fit_2d_gaussian_roi, gaussian2d_rotated, super_gaussian2d_rotated, fit_2d_gaussian_roi_NaN
+from .toolbox.fit_beam_intensity import gaussian2d_rotated, super_gaussian2d_rotated, fit_2d_gaussian_roi_NaN
 
 class GaussianFitter(QObject):
     finished = Signal(object)
@@ -34,7 +34,6 @@ class GaussianFitter(QObject):
         roi_end_row = int(np.ceil(roiPos.y() + roiSize.y()))
         roi_start_col = int(np.floor(roiPos.x()))
         roi_end_col = int(np.ceil(roiPos.x() + roiSize.x()))
-        # fit_result = fit_2d_gaussian_roi(im, roi_start_row, roi_end_row, roi_start_col, roi_end_col, function = super_gaussian2d_rotated)
         fit_result = fit_2d_gaussian_roi_NaN(im, roi_start_row, roi_end_row, roi_start_col, roi_end_col, function = super_gaussian2d_rotated)
         self.finished.emit(fit_result.best_values)
