+++ jungfrau_gui/ui_components/tem_controls/tem_controls.py
@@ -73,14 +73,12 @@ class TemControls(QGroupBox):
         self.beam_center_x.setValue(1)
         self.beam_center_x.setMaximum(globals.ncol)
         self.beam_center_x.setReadOnly(True)
-        # self.beam_center_x.valueChanged.connect(lambda value: self.spin_box_modified(self.beam_center_x))
         self.beam_center_x.editingFinished.connect(self.update_beam_center_x)
         
         self.beam_center_y = QSpinBox()
         self.beam_center_y.setValue(1)
         self.beam_center_y.setMaximum(globals.nrow)
         self.beam_center_y.setReadOnly(True)
-        # self.beam_center_y.valueChanged.connect(lambda value: self.spin_box_modified(self.beam_center_y))
         self.beam_center_y.editingFinished.connect(self.update_beam_center_y)
         
         self.label_gauss_height = QLabel()
@@ -129,8 +127,6 @@ class TemControls(QGroupBox):
             self.tem_action.enabling(False)
             self.tem_action.set_configuration()
             self.tem_action.control.draw_ellipses_on_ui.connect(self.updateFitParams)
-            # self.tem_action.control.fit_complete.connect(self.updateFitParams)
-            # self.tem_action.control.fit_complete.connect(lambda fit_result_best_values: self.updateFitParams(fit_result_best_values, draw=False))
             self.tem_action.control.remove_ellipse.connect(self.removeAxes)
             tem_section.addWidget(self.tem_stagectrl)
         else: 
@@ -184,105 +180,125 @@ class TemControls(QGroupBox):
     def reset_style(self, field):
         text_color = self.palette.color(QPalette.Text).name()
         field.setStyleSheet(f"QSpinBox {{ color: {text_color}; background-color: {self.background_color}; }}")
-
-    """ *********************************************** """
-    """ Multiprocessing Version of the gaussian fitting """
-    """ *********************************************** """
-    """ def toggle_gaussianFit(self):
-        if not self.btnBeamFocus.started:
-            if self.fitter is None:
-                self.fitter = GaussianFitterMP()
-            logging.debug(f"0.Fitter is Ready? {globals.fitterWorkerReady.value}")
-            self.fitter.updateParams(self.parent.imageItem, self.parent.roi)
-            logging.debug(f"1/2.Fitter should be Ready! Is it? --> {globals.fitterWorkerReady.value}")
-            self.fitter.finished.connect(self.updateFitParams) 
-            self.fitter.start()
-            logging.debug(f"1.Fitter is Ready? {globals.fitterWorkerReady.value}")
-            self.btnBeamFocus.setText("Stop Fitting")
-            self.btnBeamFocus.started = True  
-            # Timer started
-            self.parent.timer_fit.start(10)
-        else:
-            self.btnBeamFocus.setText("Beam Gaussian Fit")
-            self.btnBeamFocus.started = False
-            self.parent.timer_fit.stop()  
-            # Close Pop-up Window
-            self.fitter.finished.disconnect()
-            self.fitter.stop()
-            self.fitter = None
-            globals.fitterWorkerReady.value = False
-
-    def getFitParams(self):
-        self.fitter.fetch_result()
-        logging.debug(f"2.Fitter is Ready? {globals.fitterWorkerReady.value}")
-        if not globals.fitterWorkerReady.value:
-            self.fitter.updateParams(self.parent.imageItem, self.parent.roi)  
-    """
     
     """ ***************************************** """
     """ Threading Version of the gaussian fitting """
     """ ***************************************** """
-    def toggle_gaussianFit(self):
-        if not self.btnBeamFocus.started:
-            self.thread_fit = QThread()
-            self.fitter = GaussianFitter()
-            self.parent.threadWorkerPairs.append((self.thread_fit, self.fitter))                              
-            self.initializeWorker(self.thread_fit, self.fitter) # Initialize the worker thread and fitter
-            self.thread_fit.start()
-            self.fitterWorkerReady = True # Flag to indicate worker is ready
-            logging.info("Starting fitting process")
-            self.btnBeamFocus.setText("Stop Fitting")
-            self.btnBeamFocus.started = True
-            # Pop-up Window
-            if self.checkbox.isChecked():
-                self.showPlotDialog()   
-            # Timer started
-            self.parent.timer_fit.start(10)
+    def _toggle_gaussian_fit(self, btn, off_text, on_text, by_user=False):
+        """Toggle the Gaussian fitting process using the provided button and texts.
+        
+        # Args
+        btn: The button controlling the fit (e.g. self.btnBeamFocus or self.tem_tasks.btnGaussianFit).
+        off_text: The text to display when stopping the fit.
+        on_text: The text to display when starting the fit.
+        by_user (bool): Whether the user explicitly triggered the toggle.
+        """
+        if not btn.started:
+            self._start_gaussian_fit(btn, on_text, by_user)
         else:
-            self.btnBeamFocus.setText("Beam Gaussian Fit")
-            self.btnBeamFocus.started = False
-            self.parent.timer_fit.stop()  
-            # Close Pop-up Window
-            if self.plotDialog != None:
-                self.plotDialog.close()
-            self.parent.stopWorker(self.thread_fit, self.fitter)
-            self.fitter, self.thread_fit = thread_manager.reset_worker_and_thread(self.fitter, self.thread_fit)
-            self.removeAxes()
+            self._stop_gaussian_fit(btn, off_text, by_user)
+
+    def _start_gaussian_fit(self, btn, on_text, by_user=False):
+        """Starts the Gaussian fitting process."""
+        self.thread_fit = QThread()
+        self.fitter = GaussianFitter()
+        self.parent.threadWorkerPairs.append((self.thread_fit, self.fitter))
+        self.initializeWorker(self.thread_fit, self.fitter)  # Initialize worker and fitter
+        self.thread_fit.start()
+        self.fitterWorkerReady = True
+        logging.info("Starting fitting process")
+        btn.setText(on_text)
+        btn.started = True
+        if self.checkbox.isChecked():
+            self.showPlotDialog()
+        self.parent.timer_fit.start(10)
+        if by_user:
+            self.gaussian_user_forced_off = False
+
+    def _stop_gaussian_fit(self, btn, off_text, by_user=False):
+        """Stops the Gaussian fitting process."""
+        btn.setText(off_text)
+        btn.started = False
+        self.parent.timer_fit.stop()
+        if self.plotDialog is not None:
+            self.plotDialog.close()
+        self.parent.stopWorker(self.thread_fit, self.fitter)
+        self.fitter, self.thread_fit = thread_manager.reset_worker_and_thread(self.fitter, self.thread_fit)
+        self.removeAxes()
+        if by_user:
+            self.gaussian_user_forced_off = True
+
+    def toggle_gaussianFit(self):
+        """Toggle fitting using the beam focus button."""
+        # When not running, text becomes "Stop Fitting"; when stopping, revert to "Beam Gaussian Fit"
+        self._toggle_gaussian_fit(self.btnBeamFocus, off_text="Beam Gaussian Fit", on_text="Stop Fitting")
 
     def toggle_gaussianFit_beam(self, by_user=False):
-        if not self.tem_tasks.btnGaussianFit.started:
-            self.thread_fit = QThread()
-            self.fitter = GaussianFitter()
-            self.parent.threadWorkerPairs.append((self.thread_fit, self.fitter))                              
-            self.initializeWorker(self.thread_fit, self.fitter) # Initialize the worker thread and fitter
-            self.thread_fit.start()
-            self.fitterWorkerReady = True # Flag to indicate worker is ready
-            logging.info("Starting fitting process")
-            self.tem_tasks.btnGaussianFit.setText("Stop Fitting")
-            self.tem_tasks.btnGaussianFit.started = True
-            # Pop-up Window
-            if self.checkbox.isChecked():
-                self.showPlotDialog()   
-            # Timer started
-            self.parent.timer_fit.start(10)
+        """Toggle fitting using the tem_tasks button, with an optional user override."""
+        # When not running, text becomes "Stop Fitting"; when stopping, revert to "Gaussian Fit"
+        self._toggle_gaussian_fit(self.tem_tasks.btnGaussianFit, off_text="Gaussian Fit", on_text="Stop Fitting", by_user=by_user)
 
-            # If user specifically turned it ON, clear any "forced off" override
-            if by_user:
-                self.gaussian_user_forced_off = False
-        else:
-            self.tem_tasks.btnGaussianFit.setText("Gaussian Fit")
-            self.tem_tasks.btnGaussianFit.started = False
-            self.parent.timer_fit.stop()  
-            # Close Pop-up Window
-            if self.plotDialog != None:
-                self.plotDialog.close()
-            self.parent.stopWorker(self.thread_fit, self.fitter)
-            self.fitter, self.thread_fit = thread_manager.reset_worker_and_thread(self.fitter, self.thread_fit)
-            self.removeAxes()
+    # def toggle_gaussianFit(self):
+    #     if not self.btnBeamFocus.started:
+    #         self.thread_fit = QThread()
+    #         self.fitter = GaussianFitter()
+    #         self.parent.threadWorkerPairs.append((self.thread_fit, self.fitter))                              
+    #         self.initializeWorker(self.thread_fit, self.fitter) # Initialize the worker thread and fitter
+    #         self.thread_fit.start()
+    #         self.fitterWorkerReady = True # Flag to indicate worker is ready
+    #         logging.info("Starting fitting process")
+    #         self.btnBeamFocus.setText("Stop Fitting")
+    #         self.btnBeamFocus.started = True
+    #         # Pop-up Window
+    #         if self.checkbox.isChecked():
+    #             self.showPlotDialog()   
+    #         # Timer started
+    #         self.parent.timer_fit.start(10)
+    #     else:
+    #         self.btnBeamFocus.setText("Beam Gaussian Fit")
+    #         self.btnBeamFocus.started = False
+    #         self.parent.timer_fit.stop()  
+    #         # Close Pop-up Window
+    #         if self.plotDialog != None:
+    #             self.plotDialog.close()
+    #         self.parent.stopWorker(self.thread_fit, self.fitter)
+    #         self.fitter, self.thread_fit = thread_manager.reset_worker_and_thread(self.fitter, self.thread_fit)
+    #         self.removeAxes()
 
-            # If user specifically turned it OFF, set user_forced_off = True
-            if by_user:
-                self.gaussian_user_forced_off = True
+    # def toggle_gaussianFit_beam(self, by_user=False):
+    #     if not self.tem_tasks.btnGaussianFit.started:
+    #         self.thread_fit = QThread()
+    #         self.fitter = GaussianFitter()
+    #         self.parent.threadWorkerPairs.append((self.thread_fit, self.fitter))                              
+    #         self.initializeWorker(self.thread_fit, self.fitter) # Initialize the worker thread and fitter
+    #         self.thread_fit.start()
+    #         self.fitterWorkerReady = True # Flag to indicate worker is ready
+    #         logging.info("Starting fitting process")
+    #         self.tem_tasks.btnGaussianFit.setText("Stop Fitting")
+    #         self.tem_tasks.btnGaussianFit.started = True
+    #         # Pop-up Window
+    #         if self.checkbox.isChecked():
+    #             self.showPlotDialog()   
+    #         # Timer started
+    #         self.parent.timer_fit.start(10)
+
+    #         # If user specifically turned it ON, clear any "forced off" override
+    #         if by_user:
+    #             self.gaussian_user_forced_off = False
+    #     else:
+    #         self.tem_tasks.btnGaussianFit.setText("Gaussian Fit")
+    #         self.tem_tasks.btnGaussianFit.started = False
+    #         self.parent.timer_fit.stop()  
+    #         # Close Pop-up Window
+    #         if self.plotDialog != None:
+    #             self.plotDialog.close()
+    #         self.parent.stopWorker(self.thread_fit, self.fitter)
+    #         self.fitter, self.thread_fit = thread_manager.reset_worker_and_thread(self.fitter, self.thread_fit)
+    #         self.removeAxes()
+
+    #         # If user specifically turned it OFF, set user_forced_off = True
+    #         if by_user:
+    #             self.gaussian_user_forced_off = True
 
     @Slot()
     def disableGaussianFitButton(self):
@@ -292,7 +308,6 @@ class TemControls(QGroupBox):
         thread_manager.move_worker_to_thread(thread, worker)
         worker.finished.connect(self.updateFitParams)
         worker.finished.connect(self.getFitterReady)
-
 
     def getFitterReady(self):
         self.fitterWorkerReady = True
@@ -373,11 +388,13 @@ class TemControls(QGroupBox):
         # First, translate the coordinate system to the center of the ellipse,
         # then rotate around this point and finally translate back to origin.
         """ rotationTransform = QTransform().translate(xo, yo).rotate(theta_deg).translate(-xo, -yo) """
-        rotationTransform = QTransform().translate(xo, yo).rotate(-1*theta_deg).translate(-xo, -yo)
         # Create the symmetry (vertical flip) transform
         """ symmetryTransform = QTransform().translate(xo, yo).scale(1, -1).translate(-xo, -yo) """
         # Combine the rotation and symmetry transforms
         """ combinedTransform = rotationTransform * symmetryTransform  """
+
+        # All in one
+        rotationTransform = QTransform().translate(xo, yo).rotate(-1*theta_deg).translate(-xo, -yo)
 
         self.ellipse_fit.setPen(pg.mkPen('b', width=3))
         """ self.ellipse_fit.setTransform(combinedTransform) """
