+++ jungfrau_gui/ui_components/tem_controls/toolbox/fit_beam_intensity.py
@@ -197,84 +197,6 @@ def fit_2d_gaussian_roi_NaN(im, roi_start_row, roi_end_row, roi_start_col, roi_e
 
     return result_roi
 
-# @profile
-def fit_2d_gaussian_roi(im, roi_start_row, roi_end_row, roi_start_col, roi_end_col, function = gaussian2d_rotated):
-    print(f"***** roi_start_row = {roi_start_row}")    
-    print(f"***** roi_end_row = {roi_end_row}")
-    print(f"***** roi_start_col = {roi_start_col}")
-    print(f"***** roi_end_col = {roi_end_col}")
-
-    im_roi = im[roi_start_row:roi_end_row+1, roi_start_col:roi_end_col+1]
-    print(f"++++++np.max(im_roi) = {np.max(im_roi)}")
-    print(im_roi)
-    # filtered_im_roi = filter_outliers(im_roi) # remove outliers
-    # mean_intensity = np.mean(filtered_im_roi)
-    # std_intensity = np.std(filtered_im_roi)
-    # adaptive_factor = mean_intensity + 2 * std_intensity
-
-    n_columns_roi, n_rows_roi = im_roi.shape[1], im_roi.shape[0]
-
-    total_intensity = im_roi.sum()
-    
-    # Weighted average of the columns indices (xo_init)
-    col_sums = im_roi.sum(axis=0)  # Sum along the rows (column-wise sum)
-    linspace_cols = np.linspace(0, n_columns_roi-1, n_columns_roi) 
-    xo_init = np.dot(col_sums, linspace_cols) / total_intensity
-    
-    # Weighted average of the rows indices (yo_init)
-    row_sums = im_roi.sum(axis=1)  # Sum along the columns (row-wise sum)
-    linspace_rows = np.linspace(0, n_rows_roi-1, n_rows_roi) 
-    yo_init = np.dot(row_sums, linspace_rows) / total_intensity
-
-    diag_roi = np.sqrt(n_columns_roi*n_columns_roi+n_rows_roi*n_rows_roi)
-    
-    x_roi, y_roi = np.meshgrid(np.arange(n_columns_roi), np.arange(n_rows_roi))
-    z_flat_roi = im_roi.ravel()
-    x_flat_roi = x_roi.ravel()
-    y_flat_roi = y_roi.ravel()
-
-    print(f'***** x_flat_roi.shape = {x_flat_roi.shape}')
-    print(f'***** y_flat_roi.shape = {y_flat_roi.shape}')
-    print(f'***** z_flat_roi.shape = {z_flat_roi.shape}')
-
-    # Create Gaussian_2D model and parameters for ROI fitting
-    model_roi = Model(function, independent_vars=['x','y'], nan_policy='omit')
-    params_roi = Parameters()
-    # params_roi.add('amplitude', value=np.max(im_roi), min=1, max=1.2*np.max(filtered_im_roi))
-    params_roi.add('amplitude', value=0.5*np.max(im_roi), min=1, max=1.0*np.max(im_roi))
-    params_roi.add('xo', value=xo_init, min=0, max=n_columns_roi)
-    params_roi.add('yo', value=yo_init, min=0,max=n_rows_roi)
-    params_roi.add('sigma_x', value=n_columns_roi//4, min=1, max=diag_roi//2)  # Adjusted for likely ROI size
-    params_roi.add('sigma_y', value=n_rows_roi//4, min=1, max=diag_roi//2)    # Adjusted for likely ROI size
-    params_roi.add('theta', value=0, min=-np.pi/2, max=np.pi/2)
-    if function == super_gaussian2d_rotated:
-        params_roi.add('n', value=2, min=1, max=10)  # Adjust 'n' as needed
-
-    result_roi = model_roi.fit(z_flat_roi, x=x_flat_roi, y=y_flat_roi, params=params_roi)
-    fit_result = result_roi
-
-    fit_result.best_values['xo'] +=  roi_start_col+0.5
-    fit_result.best_values['yo'] +=  roi_start_row+0.5
-
-    if fit_result.best_values['sigma_x'] < fit_result.best_values['sigma_y']:
-        fit_result.best_values['sigma_x'], fit_result.best_values['sigma_y'] = fit_result.best_values['sigma_y'], fit_result.best_values['sigma_x']
-        fit_result.best_values['theta'] += np.pi/2
-
-    if fit_result.best_values['theta'] > np.pi/2:
-        fit_result.best_values['theta'] -= np.pi
-    elif fit_result.best_values['theta'] < -np.pi/2:
-        fit_result.best_values['theta'] += np.pi
-
-    fit_result.best_values['theta'] =  fit_result.best_values['theta']*180 / np.pi
-
-    return fit_result
-
-def fit_2d_gaussian_roi_fast(im, roi_coords, function = gaussian2d_rotated):
-    
-    roi_start_row, roi_end_row, roi_start_col, roi_end_col = roi_coords
-    fit_result = fit_2d_gaussian_roi(im, roi_start_row, roi_end_row, roi_start_col, roi_end_col, function = function)
-    return fit_result
-
 # Start the Qt event loop
 if __name__ == '__main__':
     format = "%(message)s"
